Fixed some issues post merge
* Size model has been changed to use userCode and not account_id. Sizes
are created when the users are created and at this stage the account_id
is not set but the userCode is set so we will use that as the FK for
sizes. This is what was causing the errors.

*Updated chat post merge

* Updated login and how the user service sets users. Michaels code was
setting this value to just the username but it should be the entire
user json file to ensure other areas of the code that use the user
service can function properly

* updated messenger post merge

* updated chat opening to use the user code and not id, code is used
when data is put into firebase

* some other minor changes to ensure consistency